requirements.txt
Flask==2.0.1
Werkzeug==2.0.3
Jinja2==3.0.3
itsdangerous==2.0.1
click==8.0.4
MarkupSafe==2.1.1
Flask-Cors==3.0.10
python-dotenv==0.19.0
supabase==1.0.3
stripe==5.4.0
requests==2.26.0
Flask-Mail==0.9.1
gunicorn==20.1.0
python-jose==3.3.0
PyJWT==2.1.0